+++ ThirdPartyAdapters/maio/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## maio Android Mediation Adapter Changelog
 
+#### Next version
+- Clean up media ID dependency in the adapter's initialize method.
+
 #### Version 2.0.7.0
 - Verified compatibility with maio SDK version 2.0.7.